+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 Under development
 =======================
 - [feature] Search data files in parent dir too (alexeevdv)
-- [feature] Basic barter system (JanSimek)
+- [feature] Basic barter system (Zervox, JanSimek)
 - [feature] Fix and update Travis build (JanSimek)
+- [feature] Load critter flags from PRO file (JanSimek)
 
 0.3.1 (2018-01-14)
 =======================